requirements.txt
flask
opencv-python
pillow
yt-dlp
